requirements.txt
streamlit
google-generativeai
pypdf
langchain
langchain-google-genai
langchain-community
faiss-cpu